create components for cashier view
