+++ front_end/src/components/Loading.tsx
@@ -5,3 +5,5 @@ import { CircularProgress } from "@mui/material";
 export const Loading: FC = () => {
   return <CircularProgress />;
 };
+
+export default Loading;
